tests: merge part tests into one file
